add modal by edit button, and function edti
