+++ src/component/mainPage.js
@@ -1,10 +1,13 @@
 // import { doc } from 'firebase/firestore';
 import {
-  logOut, createPost, recoverData, editPost,
+  logOut, createPost, editPost,
 } from '../firebase';
 // import { recoverData } from './firebase.js';
 
-function handleCreatePost() {
+const postsContainer = document.createElement('div');
+postsContainer.classList.add('postsContainer');
+postsContainer.setAttribute('id', 'postsContainer');
+const handleCreatePost = () => {
   const postContent = document.getElementById('postTextArea');
   const getUsername = localStorage.getItem('username');
   createPost(getUsername, postContent.value);
@@ -13,11 +16,7 @@ function handleCreatePost() {
   // guardar en var localStorage.getitem
   // llamar createPost(y pasar parámetros)
   // ligar handleCP al boton
-}
-
-const postsContainer = document.createElement('div');
-postsContainer.classList.add('postsContainer');
-postsContainer.setAttribute('id', 'postsContainer');
+};
 
 export const MainPage = () => {
   const principalPage = document.createElement('div'); // contiene toda la segunda vista
@@ -83,8 +82,6 @@ export const MainPage = () => {
   post.setAttribute('cols', '50');
   const publish = document.createElement('button');
   publish.addEventListener('click', handleCreatePost);
-  // publish.addEventListener('click', recoverData);
-  recoverData();
   publish.classList.add('publishPostButton');
   publishPostBtnContainer.appendChild(post);
   publishPostBtnContainer.appendChild(publish);
@@ -95,13 +92,48 @@ export const MainPage = () => {
   return principalPage;
 };
 
-export function modalEdit(oldText, idDocOrigin) {
+export const muroStructure = (doc) => {
+  const postObjects = doc.data();
+  const postUsername = postObjects.username;
+  const postTxt = postObjects.text;
+  const postLikes = postObjects.likes;
+  const idPostObject = doc.id;
+
+  const postMold = document.createElement('div'); // Contenedor del post
+  postMold.classList.add('postMold');
+  const usernameContainer = document.createElement('div'); // Contenedor username
+  usernameContainer.classList.add('usernameContainer'); // user's name
+  const username = document.createElement('h4');
+  username.classList.add('username');
+  username.innerHTML = postUsername;
+  const postContentContainer = document.createElement('div');
+  postContentContainer.classList.add('postContentContainer');
+  const postTextContent = document.createElement('h3');
+  postTextContent.classList.add('postTextContent');
+  postTextContent.innerHTML = postTxt;
+  const previousPostsLikes = document.createElement('div');
+  previousPostsLikes.classList.add('previousPostsLikes');
+  const postCountedLikes = document.createElement('h4');
+  postCountedLikes.classList.add('postCountedLikes');
+  postCountedLikes.innerHTML = `Likes: ${postLikes}`;
+  const postButtonsContainer = document.createElement('div');// donde se visualizaran los post los post
+  postButtonsContainer.classList.add('postButtonsContainer');
+  const likeBtnContainer = document.createElement('div');
+  likeBtnContainer.classList.add('likeBtnContainer');
+  const likesBtn = document.createElement('button');
+  likesBtn.classList.add('likeButton');
+  const editBtn = document.createElement('button');
+  editBtn.classList.add('editButton');
+  editBtn.setAttribute('id', 'editButton');
+  editBtn.textContent = 'Edit';
+  //
   const modalEditText = document.createElement('div');
   modalEditText.setAttribute('id', 'divModalEdit');
   const modalEditContent = document.createElement('div');
   modalEditContent.setAttribute('id', 'divModalEditContent');
   const closeModal = document.createElement('span');
   closeModal.setAttribute('id', 'spanCloseModal');
+  closeModal.textContent = 'X';
   const detail = document.createElement('h2');
   detail.setAttribute('id', 'h2Detail');
   detail.textContent = 'Edit the recipe';
@@ -110,7 +142,7 @@ export function modalEdit(oldText, idDocOrigin) {
   newPost.setAttribute('id', 'newPostTextArea');
   newPost.setAttribute('rows', '4');
   newPost.setAttribute('cols', '50');
-  newPost.placeholder = oldText;
+  newPost.placeholder = postTxt;
   const modalEditBtn = document.createElement('button');
   modalEditBtn.setAttribute('id', 'modalEditBtn');
   modalEditBtn.textContent = 'Edit';
@@ -123,9 +155,38 @@ export function modalEdit(oldText, idDocOrigin) {
 
   document.getElementById('postsContainer').appendChild(modalEditText);
 
-  const newPostContent = document.getElementById('newPostTextArea');
-  editPost(idDocOrigin, newPostContent.value);
-}
+  editBtn.addEventListener('click', () => {
+    modalEditText.style.display = 'block';
+  });
+  closeModal.addEventListener('click', () => {
+    modalEditText.style.display = 'none';
+  });
+  modalEditBtn.addEventListener('click', () => {
+    editPost(idPostObject, newPost.value);
+    modalEditText.style.display = 'none';
+  });
+  //
+  const deleteBtn = document.createElement('button');
+  deleteBtn.classList.add('deleteButton');
+  deleteBtn.textContent = 'Delete';
+
+  usernameContainer.appendChild(username);
+  postContentContainer.appendChild(postTextContent);
+  previousPostsLikes.appendChild(postCountedLikes);
+  likeBtnContainer.appendChild(likesBtn);
+  postButtonsContainer.appendChild(likeBtnContainer);
+  postButtonsContainer.appendChild(editBtn);
+  postButtonsContainer.appendChild(deleteBtn);
+
+  postMold.appendChild(usernameContainer);
+  postMold.appendChild(postContentContainer);
+  postMold.appendChild(previousPostsLikes);
+  postMold.appendChild(postButtonsContainer);
+  document.getElementById('postsContainer').appendChild(postMold);
+
+  // doc.data() is never undefined for query doc snapshots
+  console.log(doc.id, ' => ', doc.data());
+};
 // export function showPreviousPosts() {
 //   const postObjects = doc.data();
 //   const postUsername = postObjects.username;
