+++ src/firebase.js
@@ -13,7 +13,7 @@ import {
 import { async } from 'regenerator-runtime';
 // eslint-disable-next-line import/no-cycle
 import { onNavigate } from './lib/index';
-import { modalEdit } from './component/mainPage';
+import { muroStructure } from './component/mainPage';
 
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
@@ -228,63 +228,10 @@ export async function createPost(username, text) {
 export async function recoverData() {
   const querySnapshot = await getDocs(collection(dataBaseFirestore, 'publications'));
   querySnapshot.forEach((doc) => {
-    const postObjects = doc.data();
-    const postUsername = postObjects.username;
-    const postTxt = postObjects.text;
-    const postLikes = postObjects.likes;
-    const idPostObject = doc.id;
-
-    const postMold = document.createElement('div'); // Contenedor del post
-    postMold.classList.add('postMold');
-    const usernameContainer = document.createElement('div'); // Contenedor username
-    usernameContainer.classList.add('usernameContainer'); // user's name
-    const username = document.createElement('h4');
-    username.classList.add('username');
-    username.innerHTML = postUsername;
-    const postContentContainer = document.createElement('div');
-    postContentContainer.classList.add('postContentContainer');
-    const postTextContent = document.createElement('h3');
-    postTextContent.classList.add('postTextContent');
-    postTextContent.innerHTML = postTxt;
-    const previousPostsLikes = document.createElement('div');
-    previousPostsLikes.classList.add('previousPostsLikes');
-    const postCountedLikes = document.createElement('h4');
-    postCountedLikes.classList.add('postCountedLikes');
-    postCountedLikes.innerHTML = `Likes: ${postLikes}`;
-    const postButtonsContainer = document.createElement('div');// donde se visualizaran los post los post
-    postButtonsContainer.classList.add('postButtonsContainer');
-    const likeBtnContainer = document.createElement('div');
-    likeBtnContainer.classList.add('likeBtnContainer');
-    const likesBtn = document.createElement('button');
-    likesBtn.classList.add('likeButton');
-    const editBtn = document.createElement('button');
-    editBtn.classList.add('editButton');
-    editBtn.textContent = 'Edit';
-    //
-    editBtn.addEventListener('click', modalEdit(postTxt, idPostObject));
-    //
-    const deleteBtn = document.createElement('button');
-    deleteBtn.classList.add('deleteButton');
-    deleteBtn.textContent = 'Delete';
-
-    usernameContainer.appendChild(username);
-    postContentContainer.appendChild(postTextContent);
-    previousPostsLikes.appendChild(postCountedLikes);
-    likeBtnContainer.appendChild(likesBtn);
-    postButtonsContainer.appendChild(likeBtnContainer);
-    postButtonsContainer.appendChild(editBtn);
-    postButtonsContainer.appendChild(deleteBtn);
-
-    postMold.appendChild(usernameContainer);
-    postMold.appendChild(postContentContainer);
-    postMold.appendChild(previousPostsLikes);
-    postMold.appendChild(postButtonsContainer);
-    document.getElementById('postsContainer').appendChild(postMold);
-
-    // doc.data() is never undefined for query doc snapshots
-    console.log(doc.id, ' => ', doc.data());
+    muroStructure(doc);
   });
 }
+recoverData();
 
 // funcion editar texto publicacion
 export async function editPost(idDoc, newText) {
@@ -292,19 +239,4 @@ export async function editPost(idDoc, newText) {
     text: newText,
   });
 }
-
-/* export async function createPost(username, text) {
-  const postData = {
-    dateTime: new Date(),
-    likes: 0,
-    username, // cuando key y value tengan el mismo valor puedes poner el nombre y ,
-    text,
-  };
-  try{ await setDoc(publicationsAll, postData, { merge: true });
-  console.log('This value has been written to the database');
-  } catch (error){
-    console.log(`I got an error! ${error}`);
-  }
-    await updateDoc(docPublish,postData);
-} */
 // funcion eliminar texto publicacion
