+++ src/main.js
@@ -1,7 +1,7 @@
 // Este es el punto de entrada de tu aplicacion
 import { Home } from './component/home.js';
 import { MainPage } from './component/mainPage.js';
-import { createAccountFunction, loginAccountFunction, loginWithGoogle, recoverData } from './firebase.js';
+import { createAccountFunction, loginAccountFunction, loginWithGoogle } from './firebase.js';
 const rootDiv = document.getElementById('root');
 
 const routes = {
@@ -32,3 +32,4 @@ document.getElementById('LogInAccount')?.addEventListener('click', loginAccountF
 document.getElementById('googleButton')?.addEventListener('click', loginAccountFunction);
 /*const linkForgotPassword = document.getElementById('forgotPassword');
 linkForgotPassword.onclick = passwordResetEmail;*/
+
